Support for multilingual CKEditor fields
